Decouple GamesDiscoveryCategory enum class usage inside DI
