+++ app/src/main/java/com/paulrybitskyi/gamedge/ui/discovery/di/GamesDiscoveryKey.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Paul Rybitskyi, paul.rybitskyi.work@gmail.com
+ * Copyright 2021 Paul Rybitskyi, paul.rybitskyi.work@gmail.com
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,16 +14,24 @@
  * limitations under the License.
  */
 
-package com.paulrybitskyi.gamedge.commons.ui.widgets.discovery
+package com.paulrybitskyi.gamedge.ui.discovery.di
 
-import androidx.annotation.StringRes
-import com.paulrybitskyi.gamedge.commons.ui.widgets.R
+import dagger.MapKey
 
-enum class GamesDiscoveryCategory(@StringRes val titleId: Int) {
+@MapKey
+@Target(AnnotationTarget.CLASS, AnnotationTarget.FUNCTION)
+@Retention(AnnotationRetention.BINARY)
+internal annotation class GamesDiscoveryKey(val type: Type) {
 
-    POPULAR(R.string.games_category_popular),
-    RECENTLY_RELEASED(R.string.games_category_recently_released),
-    COMING_SOON(R.string.games_category_coming_soon),
-    MOST_ANTICIPATED(R.string.games_category_most_anticipated)
+
+    enum class Type {
+
+        POPULAR,
+        RECENTLY_RELEASED,
+        COMING_SOON,
+        MOST_ANTICIPATED
+
+    }
+
 
 }
